backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
mongoengine==0.27.0
pymongo==4.6.0

# Authentication
Flask-JWT-Extended==4.6.0

# Environment Variables
python-dotenv==1.0.0

# Password Hashing
bcrypt==4.1.2
